# imported/hack-govnoprod/infra/openapijson.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Curestry API",
        "description": "AI Prompt Analysis & Optimization Platform",
        "version": "0.1.0"
    },
    "paths": {
...
        "/prompt-base/add": {
            "post": {
                "tags": [
                    "prompt-base"
                ],
                "summary": "Add Prompt",
                "description": "Add a new prompt to the prompt-base.\n\nCreates a new prompt with the provided content and metadata.\nValidates the prompt structure and stores it in the database.",
                "operationId": "add_prompt_prompt_base_add_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PromptCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromptRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/prompt-base/prompts": {
            "get": {
                "tags": [
                    "prompt-base"
                ],
                "summary": "List Prompts",
                "description": "List prompts in the prompt-base with optional filtering.\n\nSupports pagination and filtering by tags, language, and format type.",
                "operationId": "list_prompts_prompt_base_prompts_get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "Number of prompts to skip",
                            "default": 0,
                            "title": "Skip"
                        },
                        "description": "Number of prompts to skip"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 1000,
                            "minimum": 1,
                            "description": "Maximum number of prompts to return",
                            "default": 100,
                            "title": "Limit"
                        },
                        "description": "Maximum number of prompts to return"
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Comma-separated list of tags to filter by",
                            "title": "Tags"
                        },
                        "description": "Comma-separated list of tags to filter by"
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Language code to filter by",
                            "title": "Language"
                        },
                        "description": "Language code to filter by"
                    },
                    {
                        "name": "format_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Format type to filter by",
                            "title": "Format Type"
                        },
                        "description": "Format type to filter by"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PromptRead"
                                    },
                                    "title": "Response List Prompts Prompt Base Prompts Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/prompt-base/prompts/{prompt_id}": {
            "get": {
                "tags": [
                    "prompt-base"
                ],
                "summary": "Get Prompt",
                "description": "Get a specific prompt by ID.",
                "operationId": "get_prompt_prompt_base_prompts__prompt_id__get",
                "parameters": [
                    {
                        "name": "prompt_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Prompt Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromptRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "prompt-base"
                ],
                "summary": "Update Prompt",
                "description": "Update an existing prompt.",
                "operationId": "update_prompt_prompt_base_prompts__prompt_id__put",
                "parameters": [
                    {
                        "name": "prompt_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Prompt Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PromptUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromptRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "prompt-base"
                ],
                "summary": "Delete Prompt",
                "description": "Delete a prompt from the prompt-base.",
                "operationId": "delete_prompt_prompt_base_prompts__prompt_id__delete",
                "parameters": [
                    {
                        "name": "prompt_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Prompt Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PromptCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Name",
                        "description": "Prompt name/title"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description",
                        "description": "Prompt description"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content",
                        "description": "The prompt text"
                    },
                    "format_type": {
                        "type": "string",
                        "maxLength": 50,
                        "title": "Format Type",
                        "description": "Prompt format (auto/xml/markdown)",
                        "default": "auto"
                    },
                    "language": {
                        "type": "string",
                        "maxLength": 10,
                        "title": "Language",
                        "description": "Prompt language",
                        "default": "en"
                    },
                    "tags": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tags",
                        "description": "Tags for categorization"
                    },
                    "extra_metadata": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Extra Metadata",
                        "description": "Additional metadata"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At",
                        "description": "Creation timestamp"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At",
                        "description": "Last update timestamp"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "content"
                ],
                "title": "PromptCreate",
                "description": "Schema for creating a new prompt."
            },
...
            "PromptRead": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Name",
                        "description": "Prompt name/title"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description",
                        "description": "Prompt description"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content",
                        "description": "The prompt text"
                    },
                    "format_type": {
                        "type": "string",
                        "maxLength": 50,
                        "title": "Format Type",
                        "description": "Prompt format (auto/xml/markdown)",
                        "default": "auto"
                    },
                    "language": {
                        "type": "string",
                        "maxLength": 10,
                        "title": "Language",
                        "description": "Prompt language",
                        "default": "en"
                    },
                    "tags": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tags",
                        "description": "Tags for categorization"
                    },
                    "extra_metadata": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Extra Metadata",
                        "description": "Additional metadata"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "id": {
                        "type": "string",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "content",
                    "created_at",
                    "updated_at",
                    "id"
                ],
                "title": "PromptRead",
                "description": "Schema for reading a prompt."
            },
...
            "PromptUpdate": {
                "properties": {
                    "name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "content": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Content"
                    },
                    "format_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Format Type"
                    },
                    "language": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Language"
                    },
                    "tags": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tags"
                    },
                    "extra_metadata": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Extra Metadata"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "title": "PromptUpdate",
                "description": "Schema for updating a prompt."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}